Fix CSS selector extraction, refactor oversized files, and fix CI errors
- Refactored `worker/pipeline/discover.ts` (oversized) into `discovery-utils.ts` to comply with the 500-line limit.
- Fixed TypeScript 'possibly undefined' error in extraction logic within `discovery-utils.ts`.
- Refactored `worker/lib/research-agent/fetcher.ts` (oversized) into modular fetchers, extractor, and transformers.
- Implemented robust CSS selector-based extraction in both discovery and research agent using `cheerio`.
- Added unit tests for HTML extraction utilities.
- Verified all changes with type checks, builds, and full unit test suite.
- Ensured compliance with repository code quality and file size policies.
